requirements.txt
langgraph>=0.2.60
langchain-core>=0.3.20
langchain-openai>=0.2.9
mcp>=1.6.0
pydantic>=2.10.0
pydantic-settings>=2.6.1
python-dotenv>=1.0.1
requests>=2.32.3
sqlalchemy>=2.0.36
psycopg[binary]>=3.2.3
google-genai>=1.33.0
streamlit>=1.39.0
